* move translations out of downloads and use servertranslate.
