+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/responses/[statusFilter]/components/DownloadSingleButton.tsx
@@ -29,27 +29,12 @@ export const DownloadSingleButton = ({
 
   const handleDownload = async () => {
     try {
-      const extraTranslations = {
-        streetAddress: {
-          en: t("addressComponents.streetAddress"),
-          fr: t("addressComponents.streetAddress"),
-        },
-        city: { en: t("addressComponents.city"), fr: t("addressComponents.city") },
-        province: { en: t("addressComponents.province"), fr: t("addressComponents.province") },
-        postalCode: {
-          en: t("addressComponents.postalCode"),
-          fr: t("addressComponents.postalCode"),
-        },
-        country: { en: t("addressComponents.country"), fr: t("addressComponents.country") },
-      };
-
       const response = (await getSubmissionsByFormat({
         formID: formId,
         ids: [responseId],
         format: DownloadFormat.HTML,
         lang: i18n.language as Language,
         revalidate: pathname.includes("new"),
-        extraTranslations,
       })) as HtmlResponse | ServerActionError;
 
       if ("error" in response) {
